:rotating_light: Rename `docs_src`'s `utils` to `validate_output`
Fixes:
```
mypy.................................................................................Failed
- hook id: mypy
- duration: 10.57s
- exit code: 2

tests/benchmarks/utils.py: error: Duplicate module named "utils" (also at
"tests/docs_src/utils.py")
tests/benchmarks/utils.py: note: See https://mypy.readthedocs.io/en/stable/running_mypy.html#mapping-file-paths-to-modules for more info
tests/benchmarks/utils.py: note: Common resolutions include: a) using `--exclude` to avoid checking one of them, b) adding `__init__.py` somewhere, c) using `--explicit-package-bases` or adjusting MYPYPATH
Found 1 error in 1 file (errors prevented further checking)
Success: no issues found in 11 source files
```
